Batch up the blob copies for a single manifest
This was the whole point of implementing the strict queue model as well
as the whole batching thing. By trying to work on the blobs for a single
image at one time, we increase the chance that we'll find good mount
candidates for future imports of the same upstream image. That seems to
bear out in my very basic real-world testing.