Add skeleton work.Detach and work.Reattach methods
While testing out some real-world copies between registries on a local
network, I noticed that it's much more common than I expected for copies
of the same blob to overlap with each other and get copied at the same
time. I guess this makes sense when certain sources and destinations
have heavy overlap and the concurrency is higher than the number of
layers in each image.

After giving some thought to various strategies, I settled on this
Context-based API for temporarily detaching a task from the concurrency
limit of its queue. My original idea was that returning an error would
somehow "defer" the current task until the next "work cycle"... and was
a complete mess in every respect. My next idea was to require passing a
Queue-specific context of some sort to each handler method, but I think
this approach is more in line with the goal of making queues more
Context-friendly, which I've wanted to do anyways.

My current idea is that a valid Detach will spawn a new worker goroutine
on the Queue, and store some state that the current worker goroutine can
read to know that it should exit after the handler returns. A valid
Reattach will reset that state and then send a token into an unbuffered
channel; each worker will check that channel before it pops from the
queue and will exit if it receives a value.