Limit to one in-flight copy per unique blob
With this, my test of copying images into two repositories in the same
registry at one time now has everything being mounted instead of copied
twice, which makes the overall operation take half as long. (This is
admittedly a pretty marginal case, but I'm glad it's working anyways.)

There's definitely some cleanup to do on this one. I like the idea of a
work.SingleFlight being directly able to detach and reattach from a
queue, and the key for the SingleFlight should perhaps include the
destination registry.