Add an initial KeyMutex unit test
Interestingly, it occurred to me while writing this test that the
channels we use to wait for the current lock holder are NOT necessary
for correctness, only to keep a KeyMutex from degrading into a spinlock.
It took me an embarrassingly long time to understand why the changes
that I thought would break my unit tests weren't actually breaking them.