Improve validation of copy specs with destination digests
There are a few obvious cases where we can know up front that a digest
in a destination spec won't be guaranteed to work, like when the digest
is unequal to the source, or the source only specifies a tag, or when we
request transforms. These will now produce errors early on instead of
potentially trying to copy a bad manifest to the registry.