Note that the project is not actively maintained
I figure I should dump this in before I re-archive the repo.

Maybe I'll come back yet again and fix some random aspect of the
implementation that's bugging me and do another drive-by unarchive and
rearchive cycle, but I provide no guarantees around that.

(The most likely manifestation of that would be if I finally build out
an interface to properly shut down a work queue. I'm actually fairly
happy with the work queue implementation and I think it could be useful
elsewhere in the future.)