Fix the deadlocks in the KeyMutex test
If we detach, we need to be able to pass our "grant" to a reattacher,
not just to a worker who can pop a key from the queue.

I'll need to think of a test that covers this case more reliably. This
is a pretty serious issue, and it's weird that I couldn't get it to show
up until I ran the tests thousands of times.