Actually launch the minimal number of worker goroutines
So, I claimed to be solving the problem of spawning more goroutines than
necessary, but in fact I made the situation far _worse_ by spooling up
the maximum allowable number of workers on every single push to the
queue.

I wonder if this has to do with why unit test failures for the previous
bugs were so hard to trigger.

I also have no idea whatsoever how to test for this. The whole concept
of the unit tests is not to understand the implementation details of how
concurrency is limited, and the _observable_ behavior really was 100%
correct.