Remove work.Queue.CloseSubmit
This actually makes me feel like all of the bugs and pain of the new
implementation were worth it. Proper cleanup is another sore point that
I've wanted to tackle for a while, but now with the new implementation
it's just not required at all. That is, adding some complexity to the
work queue implementation has now removed this potential source of
complexity for every future user.