Addendum
In the first commit on the queue branch, one of the justifications I
gave was about avoiding spurious worker startups when an existing worker
is able to take a task before a new one comes online. I don't know what
I was thinking there. The new implementation is obviously still
susceptible to the same case.

I think what I _meant_ to say was that the whole concept of "tasks"
needed to be removed, since they were ridiculous and nearly all
legitimate users of queues were just waiting on them anyways.