Clarify the bad design note
It's the smuggling of this functionality in a context.Context, not
necessarily the concept of detaching as a whole (which I still think is
a cleaner way to implement "wait for another blob copy to finish" than
any other option I've managed to come up with so far).

To be even more clear, I'm not 100% certain that using a context for
this is a bad design, but I'm leaning more strongly in that direction
than I did back when I figured cancellation would be more of a priority
to implement. Detaching and reattaching are _very_ specific to queue
handlers, and the fact that this operation is so meaningless outside of
that context hints very strongly that this needs to be represented as a
queue-specific type, and that I am a bad designer for ever considering
otherwise.