He knows when I'm awake
Okay, the naughty version was fun to write, and I enjoyed seeing how a
clear understanding of work grants could be used to justify various
implementations. But this is clearly the right way to go. First, because
simple is always better, even when it's not. Second, because this almost
certainly _is_ the common case for the work queues that Magic Mirror
actually creates, and the state checking is probably adding more
(theoretical) overhead than it (theoretically) removes (in theory,
because network I/O is orders of magnitude more expensive than whatever
is happening here, and because there's no real performance testing for
this implementation anyways).