Experiment with a new context interface for handlers
Since introducing Detach and Reattach for the work queue, I've finally
accepted that using context.Context to "connect" with the parent queue
is more of a "smuggling meaningful parameters to functions" use case
than an "optional value-add that the handler could live without" use
case. At best, this approach provides marginal value in the ability to
use KeyMutex both inside and outside of a queue, but for the most part
it just seems ripe for abuse and confusion. It doesn't feel right that
passing a generic context.Context to any function could allow it to
detach you from a work queue, and it doesn't feel right that arbitrary
functions can't know if detaching is even possible until they try it.

For emphasis: I AM ASHAMED TO HAVE EVER DESIGNED IT THIS WAY. THIS IS
EVIDENCE THAT I AM BOTH A BAD SOFTWARE ENGINEER AND A BAD PERSON. I do
not want to give the impression that I do not fully understand how
absolutely atrocious the current API is, and how OBVIOUS it should have
been to me from the moment I first wrote this code. I have been aware
that value smuggling is one of the most dangerous design elements of
context.Context since it was first introduced to Go. It is INEXCUSABLE
that I excitedly threw myself into this trap anyways.

With that out of the way… this first version is also a strange, wrong,
and bad design that I'm throwing together as part of an exploration of
the design space, i.e. so that I can understand other ways NOT to do
things and hopefully reduce the number of awful API designs I inflict on
the world in the future. The idea is that by embedding a context, you
still get "easy" access to cancellation even though detaching and
reattaching are now represented directly in the interface rather than
smuggled through a context value. Of course, doing this in an exported
interface is terrible, since of course anyone can implement that
interface and now adding methods to it is _technically_ a breaking
change. It also… just… doesn't sit right with me? It's a normal context,
yet it's not a normal context. I don't really know how to explain it, it
just leaves a bad taste in my mouth.

I think the right approach here is more like how *http.Request works in
HTTP handlers, where you have a pointer to a concrete struct, and if and
when I have an actual reason to implement contexts for handlers I can
extend the interface at that point. For now, everything related to
contexts and cancellation really needs to go away.