Continue the evil streak
This is a new spin on having detachers try to retire their work grants
immediately instead of always spinning up a new worker, which I did in
an older version of the channel implementation before removing it as the
level of complexity was completely unjustifiable.

The level of complexity is still unjustifiable, but I feel like it's
much better organized this time around. Plus, I can (hopefully?) say
that I found one of the "rare" uses of TryLock. I'm tempted to feel
worse about using TryLock than I felt about the original implementation
with a default select case, but I really don't, since I'd say they're
kind of morally equivalent to each other.