On second thought…
After sleeping on the idea of a single reattacher channel, I'm not
convinced it'll actually work. At the very least, it needs a lot more
thought than what I put in last night.

Basically, we need to maintain the invariant that work is only queued
when we can't issue enough work grants to execute it all immediately,
and the single channel (in the way I had in mind) doesn't give us a way
to guarantee that, since the communication on the channel has to be
atomic with the unlock of the work state.

The following would be a valid serialization of worker goroutines:

1. A reattacher sees that no work grants are available.
2. The reattacher gets a reference to the reattacher channel, and
   unlocks the state. It has to unlock the state before it communicates,
   otherwise the system will deadlock.
3. The reattacher is preempted before it can start to communicate on the
   channel.
4. All outstanding work grant holders complete their work and retire
   their work grants.
5. The reattacher begins to communicate on the channel, but is blocked
   as there are no work grants to transfer.

There might be some other approach based on the pattern of a buffered
channel as a semaphore, but that's probably a much larger change than
what I had in mind originally.