Remove stringkeyed.Set.ToSlice
If I wrote this from scratch today, there's a good chance I'd have used
iterators from the start, never bothering with the convenience method to
return a slice directly. Here's what that would look like.

(Interestingly, this is the first meaningful use of the optimized
Cardinality method outside of the package.)