Reattach with a single channel
A while back, I came to the realization that it's silly to have a
slice-based queue of reattacher channels, when every single Go channel
is *itself* a FIFO queue.

My original thinking was for existing workers to use a non-blocking
select to look for pending reattachers, but this didn't preclude
existing workers retiring their grants before the pending reattachers
started blocking on the channel. In other words, the reattacher has to
make itself known to existing workers through the mutex-protected work
state before it can start any channel operations, and it certainly can't
start a channel operation with the work state locked.

In retrospect, this counter-based solution seems obvious. As long as an
existing worker knows the reattacher is coming soon, it can confidently
block on the channel even if the reattacher isn't there yet.

Mutation testing showed that the decrement of the reattacher count
wasn't required for my existing unit tests to pass, so I added a new
case that carefully interposes a reattacher into a queue of unhandled
keys, ensuring that an invalid reattacher count doesn't block subsequent
handlers, and that the priority of reattachers over new keys is actually
respected.