Remove special casing for unlimited concurrency
Now, an "unlimited concurrency queue" is simply a queue that allows up
to math.MaxInt concurrent handlers. Detaching and reattaching is
meaningful regardless of the queue's concurrency limit, and touches the
count of outstanding work grants in a meaningful way.

Most importantly, this simplifies the implementation of the queue by
removing multiple special cases around key scheduling and detaching.
Whatever theoretical benefits this might have had performance-wise just
aren't worth it, especially in the context of what Magic Mirror does
(lots of requests over the Internet).

Less importantly, this cracks open the door for an interesting new
feature: dynamically changing the concurrency of an existing queue. I'm
not implementing this right now, since I don't have a clear use case for
it, but from the moment I introduced the work grant concept I worded
things very carefully to allow for exactly this possibility ("is not
lower than the maximum" rather than "is equal to the maximum"). Making
maxGrants part of the work state is another prerequisite for this
(though it also had the practical effect of telling me exactly where the
special cases for unlimited concurrency queues were, thanks to
compilation errors).