Put queued keys into a deque
I've closely reviewed this deque implementation, and I'm satisfied
enough to swap it in for my naïve slice-backed single-ended queue.

Of course, this limits reallocations of the key slice when new keys are
queued, and eliminates the wasted capacity of the sliced-off front when
new keys *aren't* being queued. But that's not why I'm doing it.

An annoying property of Magic Mirror's manifest cache is that when we
find an index and request its platform manifests, those requests get
queued behind the big batch of source requests that was dumped in at the
start of the run. We keep finding new platform manifests that we need to
copy, but it takes longer than it should to start discovering the blobs
those manifests reference.

I want something like "Get[All]Urgent" that can forcibly queue those
platform manifests ahead of all those other sources, while keeping a
single manifest cache for efficiency. I thought about doing something
special with container/heap, but a generic deque seems simpler and more
type-safe for this one use case.