Initial (untested) handling of GODEBUG=panicnil=1
See https://go.dev/play/p/u0ltkKqWmWH. I've tested the general idea
synthetically, and it produces the right values in that context, but
there should really be a unit test for this.

Note that I've carefully constructed this so t.panic is always set when
a panic occurs, and t.panicval can only be nil at runtime. It's slightly
easier to set a "t.panicnil" only when a nil panic happens, but gopls
(and presumably others) complain about the literal panic(nil) call.