Simplify task.Do a bit
After playing around with a more full-featured abstraction for this kind
of exit isolation, I came up with this simpler pattern to handle both
nil panics and Goexit. Basically, we don't need to "guess" the value of
goexit and fix it up later, but can merely flip it from true to false at
a point that can't be reached while unwinding.