Remove the async() helper from synctest cases
The entire point of this helper is to check for goroutine leaks, which
is now handled by synctest.Run panicking if all the goroutines in a
bubble are durably blocked.

That's right--we can use standard syntax for one of the language's most
important operations instead of wrapping it in some weird special thing
I made up!