Return to a more simplified handleDetach
Over time, I've gone back and forth between relatively simple and
relatively complex implementations of handleDetach. Obviously, if this
were production code rather than a pet project, I would always choose
the simplest reasonable implementation in the absence of specific
performance testing to justify a more complex approach (or at least
that's what I'm going to tell strangers on the Internet to make myself
look good). But since this _isn't_ production code, and is instead more
of a playground for me to experiment with API design and concurrency
patterns in Go, I've sometimes decided to get creative, knowing that
this is perfectly justified by the nature of the project.

My latest "creative" iteration of this was to use sync.Mutex.TryLock to
see if a detacher could quickly retire its work grant, but otherwise
fall back to spawning a new worker without a key. This was in contrast
to an implementation that always spawned a keyless worker, which might
end up doing nothing but retire the work grant.

I've recently come to the conclusion that a third approach is probably
the most sensible: always have the detacher try to retire the work grant
before spawning a new worker. The key here is that the most reasonable
use for detaching is to block on the availability of another resource,
which means that keeping the detacher from blocking isn't a useful goal.
In fact, it may be an anti-goal--if the state lock really is contended,
spawning a new goroutine just so it can _also_ block on something seems
like a waste of resources.

This seems like a strong enough justification that I might finally be
done trying different implementations of this logic, which is probably
the best reason to make the change. I suppose it's also a nice
opportunity to exercise my recent unit test changes.