Try formalizing the notion of a "SetQueue"
So it turns out you can set your "go" line to a release candidate.

I've played with some forms of a "set queue" abstraction in the past,
but this version with a generic type alias needs a lot less boilerplate
than the nominal types I've written before.

I've also gone ahead and turned work.Handler into an alias. Since it's
an internal package, and all of my own code works, the fact that this is
an API break doesn't matter.

Finally, I renamed work.NoValue to work.Empty, which I think expresses
the same concept but is much easier to read.