Optimize allocations when adding to stringkeyed.Set
This is anecdotal, but... I _think_ the fuzzer runs faster now?

Pretty obvious in retrospect: rather than creating a slice exactly large
enough to hold the current set, then appending to it (guaranteeing a
reallocation and copy), we can easily compute the length of the final
slice we actually need to sort.