Experiment with pre-computed cardinality
More memory per Set vs. less runtime on a couple of operations.

Is it worth it? It doesn't matter in the context of my use case, and I
don't have any useful synthetic tests to give a better answer.

This is another one that I want in the history to say I thought about
it, but that I don't want to keep forever.