Switch Ascii85 in stringkeyed.Set to base64
Since there's no useful performance test for this implementation,
and since it's rare that an element will require encoding at all*,
the tradeoff of more memory and comparison time for better encoding
and decoding performance hardly even warrants discussion.

What matters is that the base64 API is simpler than the Ascii85 API.

(* "Rare" in general use. In Magic Mirror it's actually impossible,
because we only use stringkeyed.Set for OCI platform references and
we pre-validate those references before putting them into the set.)