Take another pass over synctest cases
One thing I hadn't considered is cases where you previously needed a
channel for synchronization, but now can rely on synctest.Wait.

AFAIK, all of my tests really _do_ use channels in a meaningful way,
usually because I need a handler to block in order to force queueing.

In the end, the only change I made was to replace my weird use of
switching on a boolean with a plain if statement. For whatever reason I
liked the extra whitespace and lack of braces, but it would be fair for
anyone else seeing that code to call me out on it.