Migrate to synctest.Test
Go 1.25 appears set to include the first non-experimental version of
synctest, featuring the new Test function. This change updates the tests
and build tags in package work to account for that.

This means that GOEXPERIMENT=synctest on Go 1.24 will no longer activate
these tests. Until proper releases of Go 1.25 are available, gotip is
required to run them.