Try sync.WaitGroup.Go in the copier
The other sync.WaitGroup use is in work.Queue, where we add 1 to each
task's WaitGroup upon creation, and then finish it in one of the queue's
goroutines later. This is not a pattern we can reasonably port to the
new interface, though it's not like this isn't protected by the Go 1
compatibility promise. (The underlying implementation of Go is literally
the common add + go + defer done pattern, for the record. At least right
now. But probably forever given Hyrum's Law, unless it changes before
the final Go 1.25 release, no matter how much anyone thinks "f must not
panic" in the docs will make a difference.)