Implement a barebones stringkeyed.Set.UnmarshalJSONFrom
This follows the approach of the original, where we simply unmarshal
into a slice using the main JSON package, then put the elements into a
set to see if any were duplicates. This is intentionally not (yet)
taking advantage of the cool stuff we can do with the streaming decoder.