Experiment with small stringkeyed.Set additions
I don't think this is _useful_, I just wanted to see if I could do it
right.

The real problem, besides me wanting to touch this at all, is that we
always _re-encode_ the elements when we add stuff. That requires taking
a pass over every element to double-check whether it contains the magic
sequences requiring base64. A SIMD-accelerated pass, sure, but still a
pass.

What if we had a bool that said whether anything in the string was
base64'd? You'd still have to do the check on new elements, even if they
already exist elsewhere in the set. Plus, you'd have to sort and compact
the new elements separately to avoid doing that check on multiple of the
same element (again, if that's even worth caring about).

I think I'll keep this in the history but not maintain it long-term.