Try reworking pen.Result into a struct
The drawback here is that to support work.Queue's usage, we either need
the zero Result to Goexit, or we need a fake Goexit result we can borrow
as needed. Whereas my wannabe-sum-type approach both made the zero value
non-meaningful (which is actually correct here) and made it more obvious
how to use DoOrExit correctly.