Force internal/work tests into a separate package
I want an explicit test for panic(nil) handling, and I'd like a separate
package for the sake of the test-specific GODEBUG override. I've been
good about not interacting with Queue or KeyMutex internals, so this is
an easy change.

I might try the new gopls action to eliminate the dot imports, and see
what I think about the tests using the real public API with a bit of
extra verbosity on exported identifiers.