Try changing Queue.Submit* to Queue.AddNew*
"Submit" was the term I had in my head for a long time, but I'm not sure
how much I really like it. "Add" might not be perfect either; maybe I
should give in and use "Push"?

I think the "New" to help imply it only adds the new keys is useful.