Write descriptive methods for transferring work grants
In a private branch, I've been experimenting with ways to reduce the
number of raw channel operations I perform outside of well-defined
abstractions. I thought this was a good enough idea to bring over
immediately. I've always felt that the interaction on this channel was
counterintuitive, since the goroutine _sending_ its work grant has to
_receive_ on the channel where the reattachers have queued up in FIFO
order. This zero-cost abstraction makes the interaction far more
obvious.