Move task up in the file
I've been putting the "types that are an integral part of the Queue
implementation" in this zone, and I think this is really one of those.

QueueHandle can stay at the bottom; it's a user-facing API that only
makes complete sense in the context of all the implementation that comes
before it.