Support changes to a queue's concurrency limit
No, I don't really _need_ to rescale queues at runtime.

But I'm interested in the idea of removing the concurrency parameter
from NewQueue and turning it into a method call, similar to other APIs
like errgroup.Group.SetLimit. And if I go that route, I'm not interested
in an API that'll break without delicate handling, like only calling it
before anything else. I want exactly one correct way to set a queue's
concurrency limit, and I want it to work without any hidden catch.

Note that this change has gone through a few rebases, so while it has no
tests at the moment I have tried it in some capacity. Notably, my first
attempt mixed up the correct "min" with the incorrect "max", probably
because I was thinking, "the lower of these two quantities is the _max_
number of work grants we can issue."

This change is also where I first tried the function-scoped locking
pattern now on the main branch. I've elided the original version that
was written with a big critical section directly in the main function
body.

Regarding the name: "SetLimit" or "SetConcurrency" feels like the
obvious answer, but that stutters badly when you're talking about a
SetQueue (where "set" has a completely distinct meaning). Fortunately,
"limit" sounds reasonable enough as a verb, and it's not like we need to
reserve it for a getter. I think it's reasonable that if you care about
the concurrency limit that much, you can track it yourself along with
your updates to the queue.