Try to clean up a few common test patterns
Mainly the "I wish I could select a WaitGroup" pattern with done
channels. I think I was hoping I'd come up with a better abstraction for
this, but this seems good enough.

I'm also making it more clear that I'm ashamed of the unwind tests,
rather than hiding that shame in the commit message alone.