More notes on how reattachQueue works
My reading of the runtime code is that receivers are on a FIFO queue
too. At least, it's the same queue type as for senders.

There's some note in the errgroup code that [something] allows barging
iff Go channels allow barging. I don't _think_ they do right now, but
maybe someday they could (if it wouldn't break the spec and/or Hyrum's
Law)?

Taken together: enough uncertainty that I shouldn't assert FIFOness as
strongly as in the original wording.