Use function scoping for some complex critical sections
I don't like the inconsistency of my locking patterns. Some use defer,
some unlock on multiple exit paths.

Between the two, defers are obviously the preferred pattern, so I've
rewritten most use of the state lock to use them. The one exception is
tryGetQueuedKey, which itself is pretty complex. I'd like to give it
some more thought.

I'm also not following this pattern for uses of tasksMu where the
critical section is one line. I think the "traditional" pattern is more
than acceptable when the boilerplate alone would be longer.