Pull a few more changes from the rescaling branch
I cherry-picked the scoped locking work from an experimental branch
where I'm playing with the ability to dynamically adjust the concurrency
limit. I unintentionally pulled in the tryGetQueuedKey condition to deal
with a concurrency limit decrease, so I figured I might as well do the
other changes that fit into the current code: dealing with being over
the grant limit in scheduleNewKeys, and renaming maxGrants to grantLimit
since it's no longer a strict maximum.