Change "bequeath" and "obtain" to "send" and "receive"
My thought was that you're _bequeathing_ the grant because you can't do
anything useful except die once it's gone. But that's not entirely true
in the (experimental branch) case of scaling up a queue and transferring
a bunch of new work grants. Plus, the more normal terms match up better
with the doc comment.