Remove concurrency from Queue constructors
Now, queues are unlimited by default but can be limited with an explicit
call to Limit.

I was able to handle most of this rewrite with Comby. One test had a
limited queue where it didn't really matter, so I cleaned that up.
Otherwise, this was purely a mechanical change.