Improve messaging for some assertion failures
The fact that we get a line number is technically good enough, but
having something to start from on failure is better.

Also, use testify functions for all failures so we _do_ get line numbers
and consistent output formatting, and make the KeyMutex tests match some
of the newer patterns in the Queue tests.