Initial rename of package work to package parka
"Parallel Key Aggregation"

...or something like that. I'm fuzzy on what the final "a" is, and of
course the "par" violates Go's famous "concurrency is not parallelism"
aphorism. Though the Go standard library itself used to have an internal
"par" package with some stuff along these lines.

Obviously, the thinking with this rename (and all the rest of the work
on the package API) is that maybe this could be a useful package outside
of Magic Mirror.