Correctly handle Detach followed by Goexit
This is a _longstanding_ bug, that's been around basically from the
moment work.Queue started handling Goexit. parka.Map accounted for the
possibility of a handler Goexiting, but _not_ for the possibility that
the Goexited goroutine had given up its work grant.

This bug is now properly tested and fixed, following a similar (smelly)
pattern to the other unwinding tests. Naturally I'll add the panic cases
too; I just wanted this in its own commit.