Create a static encoding for function exit behaviors
Here's the promised alternative to functions in table-driven tests:
a type that so fully encapsulates every exit behavior we care about,
that you can just loop over a list of it.

Just as the swapper in the stringkeyed tests comes with a self-test,
this comes with a set of self-tests to ensure every implementation of
an exit behavior corresponds 1:1 with specific catch.Result assertions.