Add more of a package comment for parka
At some point I want to write more about this idea of decomposing a
workflow into a DAG of tasks represented by goroutines, but I don't
think this is the place for it.

I've also removed the note about there being no way to move pending keys
to the front of the queue. Originally I put that there with the thought
that I'd change InformFront, but at this point I think InformFront
should stay locked in, and this hypothetical feature deserves a separate
API (e.g. Map.Prioritize).