Terminate parka.Handle for real
Of course, for this to be correct, we must capture the detached state at
the exact moment of termination, and must never try to read it again
from then on.

I considered panicking on the termination of a terminated handler, but
there's no way to meaningfully test that code path without a second call
to Terminate, and even if that happened it might get caught like a
normal handler panic.

Of course, I still need to write a test case for the two panicking
paths, which will be extremely interesting indeed...