Improve notes about Handler + work grant concurrency
Yeah, I think it's the same old story as always. I tried to get "fancy"
and use atomics to satisfy an invariant that cannot be satisfied without
a proper mutex.

To be fair, my atomics this time were a data-race free translation of
the racy logic that existed before. I guess I could also decide that
races between Detach, Reattach, and handler exit are fatal errors that
crash the program, and somehow keep the atomics that way? Not sure
that's worth it.