Be sound in the face of concurrent Handler use
With these changes, it should be impossible to use Handle in a way that
violates a Map's work grant invariants.

For context: sync.Mutex is the size of a 64-bit pointer, so accounting
for padding this grows Handle from 24 to 32 bytes on 64-bit systems.
Locking an uncontended mutex is a single CAS, and unlocking an
uncontended mutex is a single atomic add. This really isn't bad compared
to the atomic bitfield I had before, especially when you consider that
it actually works.

To reiterate: I wrote the original atomic bitfield as a data-race free
translation of racy logic that already existed, NOT as a misguided
attempt to "optimize" something from scratch. But I'll admit that I was
_really_ hoping that translation could work as-is, and that it took me
longer than I'm proud to admit to get out of that particular rut, so I'm
making this explicit statement to point out how toxic mutex-avoidance
disease really is.