Try to decouple work grants from goroutines
Since a Map's handler is under no obligation to uphold our specific work
grant invariants (and in fact might want to break them out of spite),
I'm trying a construction that permits Handle to hold a work grant on
behalf of a handler, regardless of the number and structure of
goroutines associated with the handler.

I'm also trying to get specific in Map.work about where we loan and
retake the work grant. In doing so, I split up one deferred func into
two separate defers, so the defer chain flows more naturally up to the
loop's func literal.