Create a new torture test for Handle concurrency
I think this one is easier to read, and more likely to violate work
grant invariants if things are implemented incorrectly. In fact I'll
probably delete the old version of the test.

There seems to be another sync.WaitGroup issue in this one; it may or
may not be related to the one I just reported.