Discussion on unwinding
43a8ca57581d introduced the initial handling for panic and Goexit in
(formerly work.Queue) handlers. My commit noted that both cases screwed
up the work grant invariants, but that Goexit was more worrying since an
unrecovered panic crashes the entire program.

I think it's time for a fresh look at this.

While the work grant invariants were (and are) my main concern, I was
_also_ influenced by the "Rethinking Classical Concurrency Patterns"
talk, and the assertion that panic and Goexit ought to be forwarded out
of "internal" goroutines to the callers of methods that spawn them.
A change along these lines recently went into x/sync/errgroup after
gaining a healthy amount of support from the Go team and community...
and was subsequently reverted due to the real-world problems that come
with losing panic stacks and "hiding" panics for arbitrarily long times.
Which, if I'm being 100% honest, confirms some of my discomfort with
that change upon hearing about it for the first time.

I think there's another argument for an abstraction like errgroup.Group
or parka.Map to not recover panics: if someone _wants_ panic recovery,
they can recover() from within their own code. But if a concurrency
abstraction chooses to recover for you, you can't get the traditional
"crash the program with a useful call stack" behavior back efficiently
or consistently.

So, parka.Map definitely shouldn't recover panics. Great!

What about runtime.Goexit? It won't crash the program, so Get and
Collect must do _something_. The reasonable options are:

  1. Return zero values.
  2. Panic with a sentinel indicating the handler Goexited.
  3. Forward the Goexit.

I think 1 is clearly a poor choice, but 2 vs. 3 is harder to judge.
The only reasonable use of Goexit is in the testing package, where
FailNow is defined to Goexit the test, and is defined as only valid to
call from the goroutine running the test. If we forward Goexit, an
invalid FailNow will _probably_ exit the test in exactly the way the
author intended, even though their test is invalid per the current
testing documentation.

That pushes me toward 2: Goexit becomes a panic with a sentinel.

To summarize all the handler exit behaviors:

  - Returning from the handler... simply does what you think.
  - Panicking from the handler causes retrieval of the result to block
    forever (not a hard requirement, but I think it's a good idea).
    parka.Map does not recover the panic, so the program eventually
    terminates from an unhandled panic.
  - Goexiting from the handler causes retrieval of the result to panic
    with a well-known sentinel, indicating that the handler did not
    return but also did not crash the program.

Unless we resort to inspecting the call stack, we can't distinguish
panic(nil) from Goexit for the sake of blocking result retrieval. The
usual way to distinguish those cases is to fully recover the goroutine
and see if it continues execution, at which point we've permanently
destroyed the stack trace of the original panic. panic(nil) handling is
a neat party trick in the current version of Parka, but I won't miss it.

One last note: work() should track whether it currently has a work
grant, and defer a single `go work(nil)` for any form of exit where it's
been lost. Not useful in the panicking case, but a more robust way to
represent that invariant in general.