Panic when retrieving a Goexited handler's result
This was both easier and harder than I planned.

Easier, because the detach + Goexit test merely Informs the Goexiting
keys and doesn't try to get them. No changes there!

Harder, because the panic mock uses Goexit to replicate the unwinding
effect without crashing the program, and that test _does_ get the key
since we need to know that it blocks.