First untested draft of parka.Map.Flush
This is the API I discussed earlier, where it returns the flushed keys
in case you need to do something with them.

I'm starting to think this can't be correct unless I double-lock while
enqueueing too. I have a previously unrecognized invariant that every
unhandled task (neither in flight nor done) has a key in the queue, but
I only uphold that in Flush. Right?

Or would it be enough to switch the locking order so tasks are locked
first? That would stop any concurrent enqueue from being able to find
the tasks that will soon be missing. Right?

Every unhandled task has a key in the queue… the invariant begins with
the tasks. Once someone establishes the tasks, they just have to enqueue
the keys _eventually_. To remove tasks, we need to act _as if_ we're
dealing with the tasks first and then fixing up the key invariant, even
though we actually need the keys to know which tasks to modify.

New plan: switch the locking order to task-first.