Make tasksMu an RWMutex
The tests aren't a perfect demonstration of the improvement, but I can
see a difference when I run them tens of thousands of times. I like to
keep a healthy skepticism toward RWMutexes, but they are the "obvious"
choice for this case, and if this really is going to become a general
purpose library outside of Magic Mirror there will be more opportunities
for this to pay off.