First untested draft of parka.Map.Wait
This is intended to help prevent uncontrolled goroutine leaks when
shutting down a map early (since you'd normally just use Collect to wait
for handlers to finish).

I considered an API that returns a `<-chan struct{}` to signal doneness,
like one of the old internal/par abstractions. I'm skeptical of that API
for two reasons:

  1. It's not clear why you would `select` over this, which is the only
     benefit of a "done channel" compared to plain blocking. If you take
     a different select case, you're going to leak handlers, which is
     the entirety of what this API is meant to deal with.

  2. It limits implementation flexibility. Right now I happen to use an
     (optional) channel, but maybe a WaitGroup could be better since it
     doesn't require separate allocation?

I don't like the duplication of the closure logic across two switch
cases in tryGetQueuedKey, but I'm not sure whether I want to elevate
this to a workState-level abstraction or just clean it up within that
function. I'm leaning toward the latter, since it's the only place where
work grants are actually retired at the moment.